+++ src/useForm.ts
@@ -9,17 +9,11 @@ import shouldRenderBasedOnError from './logic/shouldRenderBasedOnError';
 import validateField from './logic/validateField';
 import assignWatchFields from './logic/assignWatchFields';
 import skipValidation from './logic/skipValidation';
-import getFieldArrayParentName from './logic/getFieldArrayParentName';
 import getFieldArrayValueByName from './logic/getFieldArrayValueByName';
-import getIsFieldsDifferent from './logic/getIsFieldsDifferent';
 import isNameInFieldArray from './logic/isNameInFieldArray';
-import isCheckBoxInput from './utils/isCheckBoxInput';
 import isEmptyObject from './utils/isEmptyObject';
-import isRadioInput from './utils/isRadioInput';
 import isSelectInput from './utils/isSelectInput';
-import isFileInput from './utils/isFileInput';
 import isObject from './utils/isObject';
-import isPrimitive from './utils/isPrimitive';
 import isFunction from './utils/isFunction';
 import isArray from './utils/isArray';
 import isString from './utils/isString';
@@ -30,7 +24,6 @@ import get from './utils/get';
 import set from './utils/set';
 import unset from './utils/unset';
 import modeChecker from './utils/validationModeChecker';
-import isMultipleSelect from './utils/isMultipleSelect';
 import unique from './utils/unique';
 import isNullOrUndefined from './utils/isNullOrUndefined';
 import isRadioOrCheckboxFunction from './utils/isRadioOrCheckbox';
@@ -39,7 +32,6 @@ import { EVENTS, UNDEFINED, VALIDATION_MODE } from './constants';
 import {
   UseFormMethods,
   FieldValues,
-  NestedValue,
   UnpackNestedValue,
   FieldName,
   InternalFieldName,
@@ -65,7 +57,8 @@ import {
   DefaultValuesAtRender,
   FlatFieldErrors,
 } from './types/form';
-import { NonUndefined, LiteralToPrimitive, DeepPartial } from './types/utils';
+import { LiteralToPrimitive, DeepPartial } from './types/utils';
+import { useFormValue } from './useFormValue';
 
 export function useForm<
   TFieldValues extends FieldValues = FieldValues,
@@ -194,135 +187,6 @@ export function useForm<
     [reRender, resolverRef],
   );
 
-  const setFieldValue = React.useCallback(
-    (
-      field: Field,
-      rawValue:
-        | FieldValue<TFieldValues>
-        | UnpackNestedValue<DeepPartial<TFieldValues>>
-        | undefined
-        | null
-        | boolean,
-    ) => {
-      const { ref, options } = field;
-      const value =
-        isWeb && isHTMLElement(ref) && isNullOrUndefined(rawValue)
-          ? ''
-          : rawValue;
-
-      if (isRadioInput(ref) && options) {
-        options.forEach(
-          ({ ref: radioRef }: { ref: HTMLInputElement }) =>
-            (radioRef.checked = radioRef.value === value),
-        );
-      } else if (isFileInput(ref)) {
-        if (isString(value)) {
-          ref.value = value;
-        } else {
-          ref.files = value as FileList;
-        }
-      } else if (isMultipleSelect(ref)) {
-        [...ref.options].forEach(
-          (selectRef) =>
-            (selectRef.selected = (value as string).includes(selectRef.value)),
-        );
-      } else if (isCheckBoxInput(ref) && options) {
-        options.length > 1
-          ? options.forEach(
-              ({ ref: checkboxRef }) =>
-                (checkboxRef.checked = (value as string).includes(
-                  checkboxRef.value,
-                )),
-            )
-          : (options[0].ref.checked = !!value);
-      } else {
-        ref.value = value;
-      }
-    },
-    [isWeb],
-  );
-
-  const setDirty = React.useCallback(
-    (name: InternalFieldName<TFieldValues>): boolean => {
-      const { isDirty, dirtyFields } = readFormStateRef.current;
-
-      if (!fieldsRef.current[name] || (!isDirty && !dirtyFields)) {
-        return false;
-      }
-
-      const isFieldDirty =
-        defaultValuesAtRenderRef.current[name] !==
-        getFieldValue(fieldsRef.current, fieldsRef.current[name]!.ref);
-      const isDirtyFieldExist = get(dirtyFieldsRef.current, name);
-      const isFieldArray = isNameInFieldArray(fieldArrayNamesRef.current, name);
-      const previousIsDirty = isDirtyRef.current;
-
-      if (isFieldDirty) {
-        set(dirtyFieldsRef.current, name, true);
-      } else {
-        unset(dirtyFieldsRef.current, name);
-      }
-
-      isDirtyRef.current =
-        (isFieldArray &&
-          getIsFieldsDifferent(
-            getFieldArrayValueByName(
-              fieldsRef.current,
-              getFieldArrayParentName(name),
-            ),
-            get(defaultValuesRef.current, getFieldArrayParentName(name)),
-          )) ||
-        !isEmptyObject(dirtyFieldsRef.current);
-
-      return (
-        (isDirty && previousIsDirty !== isDirtyRef.current) ||
-        (dirtyFields && isDirtyFieldExist !== get(dirtyFieldsRef.current, name))
-      );
-    },
-    [],
-  );
-
-  const setInternalValues = React.useCallback(
-    (
-      name: InternalFieldName<TFieldValues>,
-      value: FieldValue<TFieldValues>,
-      parentFieldName?: string,
-    ) => {
-      for (const key in value) {
-        const fieldName = `${parentFieldName || name}${
-          isArray(value) ? `[${key}]` : `.${key}`
-        }`;
-        const field = fieldsRef.current[fieldName];
-
-        if (isObject(value[key])) {
-          setInternalValues(name, value[key], fieldName);
-        }
-
-        if (field) {
-          setFieldValue(field, value[key]);
-          setDirty(fieldName);
-        }
-      }
-    },
-    [setFieldValue, setDirty],
-  );
-
-  const setInternalValue = React.useCallback(
-    (
-      name: InternalFieldName<TFieldValues>,
-      value: FieldValue<TFieldValues> | null | undefined | boolean,
-    ): boolean | void => {
-      if (fieldsRef.current[name]) {
-        setFieldValue(fieldsRef.current[name] as Field, value);
-
-        return setDirty(name);
-      } else if (!isPrimitive(value)) {
-        setInternalValues(name, value);
-      }
-    },
-    [setDirty, setFieldValue, setInternalValues],
-  );
-
   const executeValidation = React.useCallback(
     async (
       name: InternalFieldName<TFieldValues>,
@@ -434,59 +298,26 @@ export function useForm<
     return found;
   };
 
-  function setValue<
-    TFieldName extends string,
-    TFieldValue extends TFieldValues[TFieldName]
-  >(
-    name: TFieldName,
-    value: NonUndefined<TFieldValue> extends NestedValue<infer U>
-      ? U
-      : UnpackNestedValue<DeepPartial<LiteralToPrimitive<TFieldValue>>>,
-    shouldValidate?: boolean,
-  ): void;
-  function setValue<TFieldName extends keyof TFieldValues>(
-    namesWithValue: UnpackNestedValue<
-      DeepPartial<Pick<TFieldValues, TFieldName>>
-    >[],
-    shouldValidate?: boolean,
-  ): void;
-  function setValue<TFieldName extends keyof TFieldValues>(
-    names:
-      | string
-      | UnpackNestedValue<DeepPartial<Pick<TFieldValues, TFieldName>>>[],
-    valueOrShouldValidate?: unknown,
-    shouldValidate?: boolean,
-  ): void {
-    let shouldRender = false;
-    const isArrayValue = isArray(names);
-    const namesInArray = isArrayValue
-      ? (names as UnpackNestedValue<
-          DeepPartial<Pick<TFieldValues, TFieldName>>
-        >[])
-      : [names];
-
-    namesInArray.forEach((name: any) => {
-      const keyName = isString(name) ? name : Object.keys(name)[0];
-      shouldRender =
-        setInternalValue(
-          keyName,
-          isString(name)
-            ? valueOrShouldValidate
-            : (Object.values(name)[0] as any),
-        ) ||
-        isArrayValue ||
-        isFieldWatched(keyName);
-      renderWatchedInputs(keyName);
-    });
-
-    if (shouldRender || isArrayValue) {
-      reRender();
-    }
-
-    if (shouldValidate || (isArrayValue && valueOrShouldValidate)) {
-      trigger(isArrayValue ? undefined : (names as any));
-    }
-  }
+  const {
+    setFieldValue,
+    setDirty,
+    setInternalValue,
+    setValue,
+    getValues,
+  } = useFormValue({
+    isWeb,
+    fieldsRef,
+    readFormStateRef,
+    defaultValuesAtRenderRef,
+    dirtyFieldsRef,
+    fieldArrayNamesRef,
+    isDirtyRef,
+    defaultValuesRef,
+    isFieldWatched,
+    renderWatchedInputs,
+    reRender,
+    trigger,
+  });
 
   handleChangeRef.current = handleChangeRef.current
     ? handleChangeRef.current
@@ -1127,47 +958,6 @@ export function useForm<
 
     reRender();
   };
-
-  const getValue = <TFieldName extends string, TFieldValue extends unknown>(
-    name: TFieldName,
-  ): TFieldName extends keyof TFieldValues
-    ? UnpackNestedValue<TFieldValues>[TFieldName]
-    : TFieldValue =>
-    fieldsRef.current[name]
-      ? getFieldValue(fieldsRef.current, fieldsRef.current[name]!.ref)
-      : get(defaultValuesRef.current, name);
-
-  function getValues(): UnpackNestedValue<TFieldValues>;
-  function getValues<TFieldName extends string, TFieldValue extends unknown>(
-    name: TFieldName,
-  ): TFieldName extends keyof TFieldValues
-    ? UnpackNestedValue<TFieldValues>[TFieldName]
-    : TFieldValue;
-  function getValues<TFieldName extends keyof TFieldValues>(
-    names: TFieldName[],
-  ): UnpackNestedValue<Pick<TFieldValues, TFieldName>>;
-  function getValues(payload?: string | string[]): unknown {
-    const fields = fieldsRef.current;
-    if (isString(payload)) {
-      return getValue(payload);
-    }
-
-    if (isArray(payload)) {
-      return payload.reduce(
-        (previous, name) => ({
-          ...previous,
-          [name]: getValue(name),
-        }),
-        {},
-      );
-    }
-
-    const fieldValues = getFieldsValues(fields);
-
-    return isEmptyObject(fieldValues)
-      ? defaultValuesRef.current
-      : transformToNestObject(fieldValues);
-  }
 
   React.useEffect(() => {
     isUnMount.current = false;
